pc - refactor issue timeline into its own component
